Upscaling + Cropping in the HURSAT preproc
Upscaling from 0.07° to 0.25° is required to match the resolution of
ERA5. A 11x11 central cropping operation is then applied as the HURSAT
patches cover a much larger area than the extracted ERA5 patches. All of
those values are arguments to the script, so that the experiment can be
changed easily.